# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Lightweight self-hosted uptime monitoring application API.",
        "title": "Uptime Monitor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Checks whether the presented bearer token is valid. Returns 200 when valid, 401 otherwise.",
                "tags": [
                    "auth"
                ],
                "summary": "Verify API token",
                "responses": {
                    "200": {
                        "description": "Token is valid"
                    },
                    "401": {
                        "description": "Token is invalid"
                    }
                }
            }
        },
        "/api/v1/events": {
            "get": {
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Stream target updates",
                "responses": {
                    "200": {
                        "description": "Server-sent events stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/incident/{id}/read": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Mark an incident as read",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Incident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Incident marked as read"
                    },
                    "400": {
                        "description": "Invalid incident ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/incidents": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "List incidents",
                "responses": {
                    "200": {
                        "description": "List of incidents",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Incident"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/incidents/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Get latest matching incident",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target ID",
                        "name": "target_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Incident type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest matching incident",
                        "schema": {
                            "$ref": "#/definitions/database.Incident"
                        }
                    },
                    "400": {
                        "description": "Invalid target ID or incident type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Incident not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/incidents/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Mark all incidents as read",
                "responses": {
                    "204": {
                        "description": "Incidents marked as read"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/target/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the name and schedule of an existing target.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Update a target",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated target fields (name and schedule only)",
                        "name": "target",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated target",
                        "schema": {
                            "$ref": "#/definitions/database.Target"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Target not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a target and all its associated checks and incidents.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Delete a target",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Target deleted"
                    },
                    "400": {
                        "description": "Invalid target ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Target not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/targets": {
            "get": {
                "description": "Returns all monitored targets along with their most recent health checks (up to 1500 per target).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "List targets with recent checks",
                "responses": {
                    "200": {
                        "description": "List of targets with recent checks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.TargetWithChecks"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new target and schedules it for polling on the given schedule.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Create a target",
                "parameters": [
                    {
                        "description": "New target fields (name, url and schedule)",
                        "name": "target",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created target",
                        "schema": {
                            "$ref": "#/definitions/database.Target"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.Check": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "description": "CheckedAt is when the check was performed.",
                    "type": "string"
                },
                "error_message": {
                    "description": "ErrorMessage contains the error message when the check failed, if any.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier of the check.",
                    "type": "integer"
                },
                "is_up": {
                    "description": "IsUp indicates whether the target was reachable.",
                    "type": "boolean"
                },
                "response_time_ms": {
                    "description": "ResponseTimeMS is the response time in milliseconds, if any.",
                    "type": "integer"
                },
                "status_code": {
                    "description": "StatusCode is the HTTP status code returned by the endpoint, if any.",
                    "type": "integer"
                },
                "target_id": {
                    "description": "TargetID references the target this check belongs to.",
                    "type": "integer"
                }
            }
        },
        "database.Incident": {
            "type": "object",
            "properties": {
                "cause": {
                    "description": "Cause describes the reason for the incident, if known.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt is when the incident record was created.",
                    "type": "string"
                },
                "ended_at": {
                    "description": "EndedAt is when the incident ended, if it has been resolved.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier of the incident.",
                    "type": "integer"
                },
                "is_read": {
                    "description": "IsRead indicates whether the incident has been acknowledged.",
                    "type": "boolean"
                },
                "resolved": {
                    "description": "Resolved indicates whether the incident has been resolved.",
                    "type": "boolean"
                },
                "started_at": {
                    "description": "StartedAt is when the incident started.",
                    "type": "string"
                },
                "target_id": {
                    "description": "TargetID references the target this incident belongs to.",
                    "type": "integer"
                },
                "target_name": {
                    "description": "TargetName and TargetURL identify the affected item in list responses.",
                    "type": "string"
                },
                "target_url": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "Timestamp is when the incident occurred.",
                    "type": "string"
                },
                "type": {
                    "description": "Type identifies the kind of incident.",
                    "type": "string"
                }
            }
        },
        "database.Target": {
            "type": "object",
            "properties": {
                "cert_expires_at": {
                    "description": "CertExpiresAt is the NotAfter time of the TLS certificate most recently\nobserved for the target, in RFC3339 (UTC). Nil for HTTP targets or before\nthe first check observes a certificate.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt is when the target was created.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier of the target.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name is a human-readable label for the target.",
                    "type": "string"
                },
                "schedule": {
                    "description": "Schedule is the cron expression used for polling.",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt is when the target was last updated.",
                    "type": "string"
                },
                "url": {
                    "description": "URL is the endpoint that is being monitored.",
                    "type": "string"
                }
            }
        },
        "database.TargetWithChecks": {
            "type": "object",
            "properties": {
                "cert_expires_at": {
                    "description": "CertExpiresAt is the NotAfter time of the TLS certificate most recently\nobserved for the target, in RFC3339 (UTC). Nil for HTTP targets or before\nthe first check observes a certificate.",
                    "type": "string"
                },
                "checks": {
                    "description": "Checks contains the recent health check results for the target.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Check"
                    }
                },
                "created_at": {
                    "description": "CreatedAt is when the target was created.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier of the target.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name is a human-readable label for the target.",
                    "type": "string"
                },
                "schedule": {
                    "description": "Schedule is the cron expression used for polling.",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt is when the target was last updated.",
                    "type": "string"
                },
                "url": {
                    "description": "URL is the endpoint that is being monitored.",
                    "type": "string"
                }
            }
        },
        "handler.CreateTargetRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateTargetRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}